requirements.txt
fastapi
uvicorn[standard]
opencv-python
python-dotenv
google-generativeai
yt-dlp
openai